Changing SystemConfiguration to ProjectConfig and adding more variables that offer more user controls, such as Session Folder.
Fixed a bug that loading index.php (and not a specific page) would always get the Pt language.
Fixed path problem cause by recent project path redefinition, made it more generic.
Implemented a functionality to validate if the PageInstall is enabled, if it is, it will reinstall the database. Pending to update the value after the dabase is installed, so no user accidently reinstall the database by going into this page.